Move classes from OsmData to seperate classes because the endless scrolling is driving me crazy. This should have been this way from the beginning. Also add retries to routing algo without pois
