+++ lib/services/gpx_export_handler.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_map/flutter_map.dart';
 import 'package:gpx/gpx.dart';
-import 'package:hiking4nerds/services/osmdata.dart';
+import 'package:hiking4nerds/services/routing/node.dart';
 
 class GpxExportHandler{
